Get LCZ Parameters from LCZ Stewart and Oke (2012)

This function extracts 34 Local Climate Zone (LCZ) parameters based on the classification scheme developed by Stewart and Oke (2012). LCZs provide valuable information about urban morphology and are useful for urban climate and environmental studies.

getLCZparameters(
  x,
  iSelect = NULL,
  iShp = FALSE,
  iStack = FALSE,
  isave = FALSE
)

Arguments

x

A LCZ map in .tiff format SpatRaster object containing LCZ classes.

iSelect

Character vector. Specify one or more parameter names to retrieve specific parameters. For example, "SVF.min" to get the minimum Sky View Factor, or c("SVF.min", "veg.frac.max") to select multiple parameters.

iShp

Logical. If TRUE, returns all parameters as an ESRI shapefile.

iStack

Logical. If TRUE, returns all parameters as a raster stack. If FALSE, returns a list of individual parameter rasters.

isave

Set to TRUE if you want to save the your clipped resulting map as a raster.tiff or .shp file on your local machine.

Value

If 'iStack' is TRUE, it returns a raster stack containing 34 LCZ-related parameters. If 'iStack' is FALSE, it returns a list of individual parameter rasters. If 'iShp' is TRUE, it returns an ESRI shapefile containing all parameters. If 'iSelect' is specified, it returns the selected parameter(s) as a raster(s).

Examples

# Load the LCZ map
lcz_map <- raster::raster("path/to/lcz_map.tif")
#> Error in .local(.Object, ...) : 
#> 
#> Warning: GDAL support is provided by the sf and terra packages among others
#> Error in .rasterObjectFromFile(x, band = band, objecttype = "RasterLayer",     ...): Cannot create a RasterLayer object from this file. (file does not exist)

# Get LCZ parameters as a raster stack
lcz_params_stack <- getLCZparameters(lcz_map, iStack = TRUE)
#> Error in getLCZparameters(lcz_map, iStack = TRUE): object 'lcz_map' not found

# Get LCZ parameters as a list of individual parameter rasters
lcz_params_list <- getLCZparameters(lcz_map, iStack = FALSE)
#> Error in getLCZparameters(lcz_map, iStack = FALSE): object 'lcz_map' not found

# Access individual parameters from the list
specific_param <- lcz_params_list$parameter_name
#> Error in eval(expr, envir, enclos): object 'lcz_params_list' not found

# Plot a specific parameter
plot(specific_param, main = "LCZ Parameter: Parameter Name")
#> Error in plot(specific_param, main = "LCZ Parameter: Parameter Name"): object 'specific_param' not found

# Get specific parameters as a raster or shapefile
selected_params <- getLCZparameters(lcz_map, iSelect = c("SVF.min", "veg.frac.max"), iShp = TRUE)
#> Error in getLCZparameters(lcz_map, iSelect = c("SVF.min", "veg.frac.max"),     iShp = TRUE): object 'lcz_map' not found