Download the LCZ map from LCZ Generator

This function retrieves the LCZ map from LCZ Generator platform

Usage

lcz_get_map_generator(
  ID = "3110e623fbe4e73b1cde55f0e9832c4f5640ac21",
  band = "lczFilter",
  isave_map = FALSE
)

Arguments

ID

A code specifying the ID generated by LCZ Factsheet (e.g, "3110e623fbe4e73b1cde55f0e9832c4f5640ac21"). More details: https://lcz-generator.rub.de/factsheets/3110e623fbe4e73b1cde55f0e9832c4f5640ac21/3110e623fbe4e73b1cde55f0e9832c4f5640ac21_factsheet.html

band

Select the feature to use. The available options are: “lcz” and “lczFilter”. The default is "lczFilter".

isave_map

Logical. Set to TRUE if you wish to save the resulting map as a raster TIFF file on your local machine.

Value

A Terra raster TIFF file containing LCZ classes (1-17).

References

Demuzere, M., Kittner, J., Bechtel, B. (2021). LCZ Generator: a web application to create Local Climate Zone maps. Frontiers in Environmental Science 9:637455. https://doi.org/10.3389/fenvs.2021.637455

Examples

if (FALSE) { # \dontrun{
# Load the Rio de Janeiro map from LCZ Generator
my_lcz_map <- lcz_get_map_generator(ID = "3110e623fbe4e73b1cde55f0e9832c4f5640ac21")

} # }