Obtain your LCZ map from the Continental United States LCZ map

This function retrieves the Local Climate Zone (LCZ) CONUS LCZ mapping dataset created by Demuzere et al. (2020) and available at https://doi.org/10.1038/s41597-020-00605-z. It allows you to obtain the LCZ map for a specific area of interest, which can be a city, state, region, or custom-defined shape.

Usage

lcz_get_map_usa(city = NULL, roi = NULL, isave_map = FALSE, isave_usa = FALSE)

Arguments

city

A character string specifying the name of your target area based on the OpenStreetMap project.

roi

Optionally, you can provide a Region of Interest (ROI) in ESRI shapefile format to clip the LCZ map to a custom area.

isave_map

Logical. Set to TRUE if you wish to save the resulting clipped map as a raster TIFF file on your local machine.

isave_usa

Logical. Set to TRUE if you wish to save the USA LCZ map as a raster TIFF file on your local machine.

Value

A Terra raster TIFF file containing LCZ classes for the specified area of interest.

Examples

# Example 1: Load the LCZ map for a city
# my_lcz_city <- lcz_get_map_usa(city = "New York")

# Example 2: Get LCZ map for a custom region of interest
# custom_roi <- sf::st_read("custom_roi.shp")
# roi_lcz <- lcz_get_map_usa(roi = custom_roi)