Visualize the LCZ Map

This function generates a graphical representation of a Local Climate Zone (LCZ) map provided as a SpatRaster object.

Usage

lcz_plot_map(x, isave = FALSE, inclusive = FALSE, ...)

Arguments

x

A SpatRaster object containing the LCZ map to be plotted.

isave

Logical. Set to TRUE to save the plot in your working directory.

inclusive

Logical. Set to TRUE to use a colorblind-friendly palette.

...

Optional arguments to modify axis labels, legend, plot title, subtitle, and caption.

Value

A visual representation of the LCZ map in ggplot format.

See also

Examples

if (FALSE) { # \dontrun{
lcz_plot_map(LCZmap)
} # }