Visualize LCZ parameters

This function plots the parameters of an Local Climate Zone map.

Usage

lcz_plot_parameters(
  x,
  iselect = "",
  all = FALSE,
  inclusive = FALSE,
  isave = FALSE,
  save_extension = "png",
  ...
)

Arguments

x

The LCZ map in SpatRaster in a stack format.

iselect

Code vector. Specify one or more parameter:

SVFmean: Mean Sky View Factor

SVFmax: Maximum Sky View Factor

SVFmin: Minimum Sky View Factor

ARmean: Mean Aspect Ratio

ARmax: Maximum Aspect Ratio

ARmin: Minimum Aspect Ratio

BSFmean: Mean Building Surface Fraction

BSFmax: Maximum Building Surface Fraction

BSFmin: Minimum Building Surface Fraction

ISFmean: Mean Impervious Surface Fraction

ISFmax: Maximum Impervious Surface Fraction

ISFmin: Minimum Impervious Surface Fraction

PSFmean: Mean Vegetation Surface Fraction

PSFmax: Maximum Vegetation Surface Fraction

PSFmin: Minimum Vegetation Surface Fraction

TSFmean: Mean Tree Surface Fraction

TSFmax: Maximum Tree Surface Fraction

TSFmin: Minimum Tree Surface Fraction

HREmean: Mean Height Roughness Elements

HREmax: Maximum Height Roughness Elements

HREmin: Minimum Height Roughness Elements

TRCmean: Mean Terrain Roughness class

TRCmax: Maximum Terrain Roughness class

TRCmin: Minimum Terrain Roughness class

SADmean: Mean Surface Admittance

SADmax: Maximum Surface Admittance

SADmin: Minimum Surface Admittance

SALmean: Mean Surface Albedo

SALmax: Maximum Surface Albedo

SALmin: Minimum Surface Albedo

AHmean: Mean Anthropogenic Heat Output

AHmax: Maximum Anthropogenic Heat Output

AHmin: Minimum Anthropogenic Heat Output

z0: Roughness Length class

all

Logical, specifying whether to save all selected parameters into LCZ4r_output. Default is FALSE.

inclusive

Set to TRUE to a colorblind-friendly palette.

isave

Logical, indicating whether to save the plot to your directory. Default is FALSE.

save_extension

File format for saving the plot. Options: "png", "jpg", "jpeg", "tif", "pdf", "svg" (default is "png").

...

An optional modify axis, legend, and plot labels: subtitle, and caption. It does not work with title.

Value

A plot of the selected LCZ parameters in ggplot2 format

#'@references Stewart, I., and T. Oke, 2012: Local climate zones for urban temperature studies. Bull. Amer. Meteor. Soc., 93, 1879–1900. https://doi.org/10.1175/BAMS-D-11-00019.1

See also

Examples

if (FALSE) { # \dontrun{
# Plot the mean Sky View Factor (SVFmean) for your city
lcz_plot_parameters(lcz_map, iselect = "SVFmean")

# Plot multiple parameters and save them to the LCZ4r_output directory
lcz_plot_parameters(lcz_map, iselect = c("BSFmean", "AHmean"))
} # }