Analyze LCZ Time Series

This function generates a graphical representation of time series air temperature data for different Local Climate Zones (LCZs). More details: https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html

Usage

lcz_ts(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  by = NULL,
  impute = NULL,
  iplot = FALSE,
  isave = FALSE,
  palette = "Spectral",
  inclusive = FALSE,
  ylab = "Air temperature [Degree Celsius]",
  xlab = "Time",
  title = "",
  caption = ""
)

Arguments

x

A SpatRaster object containing the LCZ map. The LCZ map can be obtained using the lcz_get_map() function.

data_frame

A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.

var

Name of the variable, e.g. air temperature, in the dataframe.

station_id

Name of the station ID variable in the dataframe.

...

Utilities from selectBydata from openair package. A start date string in the form e.g. “1/2/1999” or in format i.e. “YYYY-mm-dd”, “1999-02-01”. A year or years to select e.g. year = 1998:2004 to select 1998-2004 inclusive or year = c(1998, 2004) to select 1998 and 2004. A month or months to select. Can either be numeric e.g. month = 1:6 to select months 1-6 (January to June), or by name e.g. month = c(“January”, “December”).

time.freq

Defines the time period to average to. Default is “hour”, but includes “day”, “week”, “month” or “year”.

by

data frame time-serie split: “year”, “season”, “seasonyear”, “month”, “monthyear”, “weekday”, “weekend”, “site”, “daylight”, “dst” (daylight saving time).See argument type in openair package: https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option

impute

Method to impute missing values (“mean”, “median”, “knn”, “bag”).

iplot

Set to TRUE if you want to save the plot in your working directory.

isave

Save the plot into your directory.

palette

Define your qualitative color palette. Default is "Spectral". There are more 100 color-palettes grDevices::hcl.pals()

inclusive

Set to TRUE to a colorblind-friendly palette.

ylab

y-axis name.

xlab

y-axis name. Default is “Time”

title

y-axis name. Default is “" "”.

caption

source data. Default can be “Source:Stewart and Oke, 2012; Demuzere et al.2022."”.

Value

A visual representation of the time series of air temperature of LCZ in ggplot format

See also

Examples

# Hourly air temperature values in 2019.
# my_ts <- lcz_ts(lcz_map, df = lcz_data, var = "airT",
#                 station_id = "station", year = 2019)