Explore LCZ Thermal Anomalies

This function generates a graphical representation of thermal anomaly for different Local Climate Zones (LCZs)

Usage

lcz_anomaly(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  by = NULL,
  hemisphere = "northern",
  impute = NULL,
  iplot = FALSE,
  isave = FALSE,
  inclusive = FALSE,
  ylab = "Thermal Anomaly [Degree Celsius]",
  xlab = "Station",
  title = "",
  caption = "Source: LCZ4r; Stewart and Oke, 2012; Demuzere et al.2022."
)

Arguments

x

A SpatRaster object containing the LCZ map. The LCZ map can be obtained using the lcz_get_map() function.

data_frame

A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.

var

Name of the variable, e.g. air temperature, in the dataframe.

station_id

Name of the station ID variable in the dataframe.

...

Utilities from selectBydata from openair package. A start date string in the form e.g. “1/2/1999” or in format i.e. “YYYY-mm-dd”, “1999-02-01”. A year or years to select e.g. year = 1998:2004 to select 1998-2004 inclusive or year = c(1998, 2004) to select 1998 and 2004. A month or months to select. Can either be numeric e.g. month = 1:6 to select months 1-6 (January to June), or by name e.g. month = c(“January”, “December”).

time.freq

Defines the time period to average to. Default is “hour”.

by

Temporal components of the data (e.g.,“sec”, “min”, “hour”, “day”, “DSTday”, “week”, “month”, “quarter” or “year”). See argument type in openair package.

hemisphere

Hemisphere for splitting data into seasons (“northern” or “southern”).

impute

Method to impute missing values (“mean”, “median”, “knn”, “bag”).

iplot

Set to TRUE if you want to save the plot in your working directory.

isave

Save the plot into your directory.

inclusive

Set to TRUE to a colorblind-friendly palette.

ylab

y-axis name. Default is “Thermal Anomaly Degree Celsius”.

xlab

y-axis name. Default is “Station”

title

y-axis name. Default is “" "”.

caption

source data. Default is “Source: LCZ4r, Stewart and Oke, 2012; Demuzere et al.2022."”.

Value

A visual representation of the time series of air temperature of LCZ in ggplot format

See also

Examples

# Hourly air temperature values in 2019.
# my_ts <- lcz_anomaly(my_map, df = air_UCON, var = "airT", station_id = "station", year = 2019)