Visualize the interpolated LCZ-Thermal Anomaly map

This function generates a spatial interpolation of thermal anomaly for different Local Climate Zones (LCZs). More details: https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html

Usage

lcz_anomaly_map(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  sp.res = 100,
  tp.res = "hour",
  vg.model = "Sph",
  by = NULL,
  impute = NULL,
  isave = FALSE,
  LCZinterp = TRUE
)

Arguments

x

A SpatRaster object containing the LCZ map. The LCZ map can be obtained using the lcz_get_map() function.

data_frame

A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.

var

Name of the variable for interpolation, e.g. air temperature, in the dataframe.

station_id

Name of the station ID variable in the dataframe.

...

Additional arguments for the selectBydata from openair package. These include:

A start date string in the form "1/2/1999" or in format "YYYY-mm-dd", e.g., "1999-02-01".

A year or years to select, e.g., year = 1998:2004 to select 1998-2004 inclusive, or year = c(1998, 2004) to select 1998 and 2004.

A month or months to select. Can either be numeric, e.g., month = 1:6 to select January to June, or by name, e.g., month = c("January", "December").

sp.res

Spatial resolution in unit of meters for interpolation. Default is 100.

tp.res

Temporal resolution, the time period to average to. Default is “hour”, but includes “day”, “week”, “month” or “year”.

vg.model

If kriging is selected, the list of viogrammodels that will be tested and interpolated with kriging. Default is "Sph". The model are "Sph", "Exp", "Gau", "Ste". They names respective shperical, exponential, gaussian,Matern familiy, Matern, M. Stein's parameterization.

by

data frame time-serie split: “year”, “season”, “seasonyear”, “month”, “monthyear”, “weekday”, “weekend”, “site”, “daylight”(daytime and nighttime).See argument type in openair package: https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option

impute

Method to impute missing values (“mean”, “median”, “knn”, “bag”).

isave

Save the map into your directory.

LCZinterp

If set to TRUE (default), the LCZ interpolation approach is used. If set to FALSE, conventional interpolation without LCZ is used.

Value

A map of LCZ-thermal anomalies in terra raster GeoTIF format

See also

Examples

if (FALSE) {
 #Daily air temperature values in September 2019 .
 my_anomaly <- lcz_anomaly_map(lcz_map, data_frame = lcz_data, var = "airT",
                              station_id = "station", tp.res = "day", sp.res= "100",
                               year = 2019, month=9)
 }